requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cryptography==3.0
dominate==2.6.0
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Script==2.0.6
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==5.1.0
pyOpenSSL==19.1.0
requests==2.25.1
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.26.2
visitor==0.1.3
Werkzeug==1.0.1
yarl==1.6.3
